Moved exam assignment functionality to dashboard (using mixins)
