+++ dashboard/dashboard_links.py
@@ -14,7 +14,8 @@ class DashboardLinks():
         """
         return [
             ("Übersicht", cls.get_index_link()),
-            ("Ersties", reverse_lazy('dashboard:students:index'))
+            ("Ersties", reverse_lazy('dashboard:students:index')),
+            ("Klausur", reverse_lazy('dashboard:exam:assignment_new')),
         ]
 
     @classmethod
