+++ dashboard/urls.py
@@ -7,4 +7,5 @@ urlpatterns = patterns('',
     url(r'^$', views.IndexView.as_view(), name='index'),
     url(r'^permissionmissing$', views.PermissionMissingView.as_view(), name='missing_permission'),
     url(r'^students/', include('students.dashboard_urls', namespace='students')),
+    url(r'^exam/', include('exam.dashboard_urls', namespace='exam')),
 )
