+++ exam/templates/exam/assignment_form.html
@@ -2,8 +2,9 @@
 {% block title %}Neue Zuteilung | Klausur{% endblock %}
 
 {% block content_dashboard_view %}
-<div id="mainForm">
+<h3>Klausurzuteilung durchführen</h3>
 
+<div id="mainForm">
     <table class="table">
         <tr><th>Angemeldete Ersties (alle Gruppen):</th><th>{{ student_count }}</th></tr>
         <tr class="{% if free_places_1 > student_count %}success{% else %}danger{% endif %}"><td>Verfügbare Plätze (1 Platz Abstand):</td><td>{{ free_places_1 }}</td></tr>
