+++ exam/templates/exam/success.html
@@ -3,8 +3,9 @@
 {% block title%}Zuteilung durchgeführt | Klausurzuteilung{% endblock %}
 
 {% block content_dashboard_view %}
+<h3>Klausurzuteilung durchführen</h3>
 
-    <div class="alert alert-success">{% bootstrap_icon "ok" %} Zuteilung durchgeführt.</div>
-    <a href="{%  url 'exam:assignment_new' %}" class="btn btn-default">Neue Zuteilung durchführen</a>
+<div class="alert alert-success">{% bootstrap_icon "ok" %} Zuteilung durchgeführt.</div>
+<a href="{%  url 'dashboard:exam:assignment_new' %}" class="btn btn-default">Neue Zuteilung durchführen</a>
 
 {% endblock %}
